Fix playback of large recorded mkv files.
CHTSPVFS::SendFileSeek() was returning an error due to 32-bit ssize_t overflow which caused matroska header parsing to fail.

+++ addons/pvr.tvh/src/Tvheadend.h
@@ -314,7 +314,7 @@ private:
 
   bool      SendFileOpen  ( bool force = false );
   void      SendFileClose ( void );
-  ssize_t   SendFileSeek  ( int64_t pos, int whence, bool force = false );
+  long long SendFileSeek  ( int64_t pos, int whence, bool force = false );
 };
 
 /*
